TestTask/requirements.txt
Django==4.0.5
django-model-utils==4.2.0
django-ranged-response==0.2.0
djangorestframework==3.13.1
